Change chocolatey install to recommended commands
Chocolatey install failed on first run on new machine so try changing it to match the current recommended approach.
Also changed the elevation to admin permissions as that didn't seem to work anyway and just bomb out if not run as administrator.
Added set of execution policy to allow for running of local powershell scripts and remote scripts as long as they are properly signed.
Try installing code settings VS Code extension.