Turn off SnapAssist
When snapping a window to one side of the screen this gives you the choice of what you want to snap to the other side.
I find that irritating as I quite often have a few windows snapped to either side and I have to exit from the selection more than I use it.