Improve setup of WSL
VirtualMachinePlatform also needs to be enabled so add that but only require a single reboot even if both need to be enabled.
Also moved WSL setup to much earlier in the script as it may require a reboot so script will need to be re-run and don't want to have to process loads of packages multiple times.